requirements.txt
pyautogen
chess
groq
matplotlib
numpy
pandas
yfinance
python-dotenv

